DM-188 Moved out SleeperSearch child components to separate files
