+++ frontend/src/feature/search/components/SleeperSearch.tsx
@@ -1,42 +1,18 @@
 // -------------------- Imports --------------------
 import {
-  Box,
-  Button,
-  CircularProgress,
-  FormControl,
-  FormControlLabel,
-  Paper,
-  Radio,
-  RadioGroup,
-  Snackbar,
   Stack,
-  TextField,
   Typography,
 } from "@mui/material";
 
-import { useRouter } from "@tanstack/react-router";
-import { useSuspenseQuery } from "@tanstack/react-query";
-
-import { Route as LeagueRoute } from "@routes/leagues.$leagueId";
-import { DisplayLeaguesList } from "@components/DisplayLeaguesList";
-import SelectSeasonDropDown from "@components/SelectSeasonDropDown";
-
-import useGetUserLeaguesSleeper from "@feature/search/hooks/useGetUserLeaguesSleeper";
 import useSleeperSearchParams from "@feature/search/hooks/useSleeperSearchParams";
-import useSaveSleeperLeague from "@feature/leagues/hooks/useSaveTeam";
-
-import { useGetSavedLeagues } from "@hooks/useGetSavedLeagues";
-import { useNotification } from "@hooks/useNotification";
-import { useAppSelector } from "@app/hooks";
-
-import useDeleteLeague from "../hooks/useDeleteLeague";
-import { sleeper_getUser } from "@services/sleeper";
+import SleeperLeaguesList from "./SleeperLeaguesList";
+import SleeperSearchForm from "./SleeperSearchForm";
 
 // -------------------- Types --------------------
 /**
  * Props shared between SleeperSearch sub-components.
  */
-type SleeperSearchComponentProps = {
+export type SleeperSearchComponentProps = {
   searchType: string;
   season: string;
   searchText: string;
@@ -88,7 +64,7 @@ export default function SleeperSearch({ season: initSeason = '2025', searchType:
       </Typography>
 
       {showLeagues ?
-        <SleeperDisplayUsernameSearchResult
+        <SleeperLeaguesList
           searchType={searchType}
           season={season}
           searchText={searchText}
@@ -100,7 +76,7 @@ export default function SleeperSearch({ season: initSeason = '2025', searchType:
           setParamsFalse={setParamsFalse}
         />
         :
-        <SleeperFindLeague
+        <SleeperSearchForm
           searchType={searchType}
           season={season}
           searchText={searchText}
@@ -114,312 +90,5 @@ export default function SleeperSearch({ season: initSeason = '2025', searchType:
         />
       }
     </Stack>
-  );
-}
-
-// -------------------- Child Components --------------------
-/**
- * A form component for searching Sleeper leagues by **Username** or **League ID**.
- *
- * - If searching by `League ID`, it will attempt direct navigation to the league.
- * - If searching by `Username`, the user must also select a season.
- *
- * @param props - {@link SleeperSearchComponentProps}
- * @returns The rendered league search form.
- */
-function SleeperFindLeague({
-  searchType,
-  season,
-  searchText,
-  handleTextChange,
-  setSeason,
-  handleSearchTypeChange,
-  checkValidParams,
-  handleLeagueSearch,
-}: SleeperSearchComponentProps) {
-  const { showSuccess, showError } = useNotification();
-
-  /**
-   * Handles form submission and executes the correct search action.
-   *
-   * @param e - The form submit event.
-   */
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (searchType === "League ID" && searchText !== "") {
-      const success = await handleLeagueSearch!();
-      if (success) {
-        showSuccess("Navigating to League");
-      } else {
-        showError("League not found");
-      }
-    } else {
-      checkValidParams();
-    }
-  };
-
-  return (
-    <Paper
-      elevation={3}
-      sx={{
-        borderRadius: 3,
-        p: 4,
-        m: 2,
-        width: "90%",
-        maxWidth: 400,
-        mx: "auto",
-        border: "1px solid",
-        borderColor: "divider",
-      }}
-    >
-      {/* Header */}
-      <Box sx={{ textAlign: "center", mb: 3 }}>
-        <Typography variant="h4" component="h2" gutterBottom color="primary">
-          Find League
-        </Typography>
-        <Typography variant="body2" color="text.main">
-          Search by username or league ID
-        </Typography>
-      </Box>
-
-      {/* Form */}
-      <FormControl fullWidth>
-        {/* Radio Selection */}
-        <RadioGroup
-          row
-          value={searchType}
-          onChange={handleSearchTypeChange}
-          sx={{
-            mb: 3,
-            justifyContent: "center",
-            "& .MuiFormControlLabel-root": { mx: 2 },
-          }}
-        >
-          <FormControlLabel
-            value="Username"
-            control={<Radio />}
-            label="Username"
-            sx={{ "& .MuiFormControlLabel-label": { fontWeight: 500 } }}
-          />
-          <FormControlLabel
-            value="League ID"
-            control={<Radio />}
-            label="League ID"
-            sx={{ "& .MuiFormControlLabel-label": { fontWeight: 500 } }}
-          />
-        </RadioGroup>
-
-        {/* Input Fields */}
-        <Box sx={{ mb: 3 }} display="flex" gap={2}>
-          <TextField
-            label={searchType}
-            required
-            variant="outlined"
-            onChange={handleTextChange}
-            value={searchText}
-            sx={{
-              flex: 2,
-              "& .MuiOutlinedInput-root": { borderRadius: 2 },
-            }}
-          />
-          {searchType === "Username" && (
-            <Box sx={{ flex: 1 }}>
-              <SelectSeasonDropDown
-                updateSeason={setSeason}
-                selectedYear={season}
-                label_name="Year"
-                disabled={false}
-              />
-            </Box>
-          )}
-        </Box>
-
-        {/* Submit Button */}
-        <Button
-          onClick={handleSubmit}
-          variant="contained"
-          size="large"
-          fullWidth
-          sx={{
-            py: 1.5,
-            borderRadius: 2,
-            transition: "all 0.3s ease",
-            "&:hover": { transform: "translateY(-2px)" },
-          }}
-        >
-          <Typography variant="button" color="primary.contrastText">
-            Search League
-          </Typography>
-        </Button>
-      </FormControl>
-    </Paper>
-  );
-}
-
-/**
- * Displays a list of Sleeper leagues that match the current search criteria.
- *
- * @param props - {@link SleeperSearchComponentProps}
- * @returns The rendered leagues list with options to save/delete.
- */
-function SleeperDisplayUsernameSearchResult({
-  searchType,
-  season,
-  searchText,
-  setSeason,
-  setParamsFalse: back,
-}: SleeperSearchComponentProps) {
-  const username = useAppSelector((state) => state.authReducer.username);
-
-  const { leagues, loading, error } = useGetUserLeaguesSleeper(searchText, season);
-  const { data: user_id } = useSuspenseQuery({
-    queryKey: ["sleeper_user", searchText],
-    queryFn: () => sleeper_getUser(searchText),
-    select: (data) => data?.user_id,
-  });
-  const { data: userLeagues } = useGetSavedLeagues();
-
-  const deleteLeague = useDeleteLeague();
-  const saveTeamMutate = useSaveSleeperLeague();
-
-  const sleeperLeagues = userLeagues?.reduce<string[]>((result, league) => {
-    if (league.platform === "sleeper") result.push(league.league_id);
-    return result;
-  }, []);
-
-  const router = useRouter();
-
-  /**
-   * Navigates the user to a given league details page.
-   *
-   * @param id - The league ID to navigate to.
-   */
-  const handleNavigateToLeague = (id: string) => {
-    router.navigate({
-      to: LeagueRoute.to,
-      params: { leagueId: id },
-    });
-  };
-
-  /**
-   * Saves a league for the logged-in user.
-   *
-   * @param league_id - The league ID to save.
-   * @returns Whether the save was successful.
-   */
-  const saveLeague = async (league_id: string) => {
-    try {
-      saveTeamMutate.mutate({ league_id, user_id: user_id! });
-      return saveTeamMutate.isSuccess;
-    } catch {
-      return false;
-    }
-  };
-
-  /**
-   * Deletes a league from the logged-in user's saved list.
-   *
-   * @param league_id - The league ID to delete.
-   * @returns Whether the delete was successful.
-   */
-  const handleDeleteLeague = async (league_id: string) => {
-    deleteLeague.mutate({ platform: "sleeper", league_id });
-    return deleteLeague.isSuccess;
-  };
-
-  if (loading) return <CircularProgress />;
-  if (error) {
-    back();
-    return <Snackbar open={!!error} message={error} />;
-  }
-
-  return (
-    <Paper
-      elevation={3}
-      sx={{
-        borderRadius: 3,
-        m: 2,
-        maxWidth: 800,
-        mx: "auto",
-        border: "1px solid",
-        borderColor: "divider",
-      }}
-    >
-      {/* Search/Filter Section */}
-      <Box
-        display="flex"
-        alignItems="center"
-        sx={{
-          p: 3,
-          gap: 2,
-          borderBottom: "1px solid",
-          borderColor: "divider",
-        }}
-      >
-        <Button
-          variant="outlined"
-          onClick={back}
-          color="primary"
-          sx={{
-            height: "56px",
-            borderRadius: 2,
-            textTransform: "none",
-            px: 3,
-            borderColor: "primary.main",
-            transition: "all 0.3s ease",
-            "&:hover": {
-              borderColor: "primary.dark",
-              backgroundColor: "primary.main",
-              color: "primary.contrastText",
-              transform: "translateY(-1px)",
-              boxShadow: "0 2px 8px primary.light",
-            },
-          }}
-        >
-          <Typography variant="body1">Back</Typography>
-        </Button>
-
-        <Box display="flex" gap={2} sx={{ flex: 1 }}>
-          <TextField disabled label={searchType} value={searchText} />
-          <FormControl sx={{ flex: 1, minWidth: 150 }}>
-            <SelectSeasonDropDown
-              updateSeason={setSeason}
-              selectedYear={season}
-              label_name="Change Year"
-              width={150}
-            />
-          </FormControl>
-        </Box>
-      </Box>
-
-      {/* Results Section */}
-      {!leagues || leagues.length === 0 ? (
-        <Box
-          sx={{
-            p: 4,
-            textAlign: "center",
-            backgroundColor: "background.default",
-            color: "text.primary",
-          }}
-        >
-          <Typography variant="body1">No Leagues Found</Typography>
-        </Box>
-      ) : (
-        <Box sx={{ p: 2 }}>
-          <DisplayLeaguesList
-            onLeagueClick={handleNavigateToLeague}
-            displayAvatar={true}
-            leagues={leagues}
-            loggedIn={!!username}
-            saveDelete={{
-              saveLeague,
-              userLeagues: sleeperLeagues || [],
-              deleteLeague: handleDeleteLeague,
-            }}
-          />
-        </Box>
-      )}
-    </Paper>
   );
 }
